Strengthen MCP policy example
