+++ examples/05-mcp-tool-call/requirements.txt
@@ -1,4 +1,3 @@
-jamjet>=0.1.2
-openai>=1.0
+jamjet>=0.8.0
 pydantic>=2.0
 mcp>=1.12.4
