+++ examples/05-mcp-tool-call/setup.sh
@@ -29,7 +29,4 @@ echo ""
 echo "Done. To run:"
 echo ""
 echo "  source $ACTIVATE_SCRIPT"
-echo "  export OPENAI_API_KEY=ollama"
-echo "  export OPENAI_BASE_URL=http://localhost:11434/v1"
-echo "  export MODEL_NAME=llama3.2:3b"
 echo "  python main.py"
